+++ tests/puppeteer/pages/BO/shopParameters/productSettings/index.js
@@ -77,7 +77,7 @@ module.exports = class productSettings extends BOBasePage {
   /**
    * Enable/Disable force update of friendly URL
    * @param toEnable
-   * @returns {Promise<string|*>}
+   * @returns {Promise<string>}
    */
   async setForceUpdateFriendlyURL(toEnable = true) {
     await this.waitForSelectorAndClick(this.switchForceUpdateFriendlyURLLabel.replace('%TOGGLE', toEnable ? 1 : 0));
